Moved _preprocess.
